Time: 33 ms (9.09%), Space: 11.4 MB (18.18%) - LeetHub

+++ 3768-check-if-digits-are-equal-in-string-after-operations-i/3768-check-if-digits-are-equal-in-string-after-operations-i.cpp
@@ -3,7 +3,7 @@ public:
     bool hasSameDigits(string s) {
         while (s.length() > 2) {
             string new_s = "";
-            for (size_t i = 0; i < s.length() - 1; ++i) {
+            for (int i = 0; i < s.length() - 1; ++i) {
                 int new_digit = (s[i] - '0' + s[i + 1] - '0') % 10;
                 new_s += to_string(new_digit);
             }
